refactor(agent): extract execution domain from loop
