+++ packages/agent/src/execution/file-lock.ts
