+++ packages/agent/src/loop/index.ts
@@ -1,6 +1,4 @@
-// Loop module — conversation/execution supervisors, DAG execution, file locking
-
-export { FileLock } from "./file-lock";
+// Loop module — conversation supervisor
 
 export {
   ConversationSupervisor,
@@ -13,41 +11,3 @@ export {
   type ConversationHistory,
   type ConversationSupervisorResult,
 } from "./conversation-supervisor";
-
-export type {
-  SupervisorDecision,
-  SummarizedHistory,
-  ExecutionPlan,
-  ExecutionPlanStep,
-  ExecutionSupervisorConfig,
-  ExecutionSupervisorResult,
-  StepOutcome,
-  DispatchContext,
-  DispatchExecutionInput,
-  DispatchOutput,
-  DispatchReviewDecision,
-  DispatchReviewInput,
-  DispatchTask,
-} from "./execution-types";
-
-export { ExecutionSupervisor } from "./execution-supervisor";
-
-export {
-  buildDependencyGraph,
-  completeTaskAndUnblockDependents,
-} from "./execution-graph";
-
-export {
-  resolveDispatchHybridRuntime,
-  assignAgentsToReadyTasks,
-  resolveFallbackAgentAssignment,
-  resolveWorkerRuntimeForTask,
-} from "./execution-assignment";
-
-export {
-  decideFailedStepAction,
-  reviewTaskResult,
-  sendReviewFeedback,
-  requestHandoffDocument,
-  rotateAgent,
-} from "./execution-review";
